UPDATE directory structure REFACTORING move tool's utils into tools dir.
