+++ docs/.vuepress/tools/jsdoc-convert/jsdoc.js
@@ -4,7 +4,8 @@ const jsdoc2md = require('jsdoc-to-markdown'); // eslint-disable-line import/no-
 const dmd = require('dmd'); // eslint-disable-line import/no-unresolved
 const path = require('path');
 const fs = require('fs');
-const { logger } = require('../../utils');
+
+const { logger } = require('../utils');
 
 /// parameters
 const pathToSource = '../../../../src';
@@ -265,5 +266,5 @@ if (errors.length) {
   process.exit(1);
 }
 
-logger.log('OK!');
+logger.success('OK!');
 process.exit(0);
