+++ docs/README.md
@@ -38,19 +38,24 @@ npm run docs:start
 ## Directory structure
 
 ```
-- /next                     # The next version of documentation, unavailable on production build.
-    + /api                  # Automatically generated files from JsDoc. Do not edit!
-    *.md                    # The content of documentation
-+ /SEMVER.NUMBER            # Version of documentation, for instance 8.4, 8.5
-- /.vuepress                # Vuepress files and configuration
-    + /public               # Public assets
-    + /theme                # Theme overwrites and customizations
+- /next                         # The next version of documentation, unavailable on production build.
+    + /api                      # Automatically generated files from JsDoc. Do not edit!
+    *.md                        # The content of documentation
++ /SEMVER.NUMBER                # Version of documentation, for instance 8.4, 8.5
+- /.vuepress                    # Vuepress files and configuration
+    - /containers               # Markdown containers `:::`
+        + /examples             # Container for running code snippets `::: example #exampleId .class :preset --html 1 --js 2`
+        sourceCodeLink.js       # Add link "Source Code" on the right, `::: source-code-link [URL]`
+    + /handsontable-manager     # This module provides api to run any example in any handsontable version with selected preset.
+    + /public                   # Public assets
+    + /theme                    # Theme overwrites and customizations
     - /tools            
-        + /jsdoc-convert    # Tool used for converting jsdoc to markdown
-        + /version          # Tool used for rollup a new version
-    config.js               # Configuration file
-    examples.js             # Handle live code preview and jsfiddle button
-    helpers.js              # Common helpers to get sidebars and list of versions
-    highlight.js            # Code highlight configuration
+        + /jsdoc-convert        # Tool used for converting jsdoc to markdown
+        + /version              # Tool used for rollup a new version
+        check-links.js          # Tool which check all links in the documentation
+        utils.js                # Utilities for tools
+    config.js                   # Configuration file
+    helpers.js                  # Common helpers to get sidebars and list of versions
+    highlight.js                # Code highlight configuration
 ```
